rename github to github api folder
