+++ src/networks/githubApi/checkMembershipForUser.ts
